requirements.txt
pyspark==3.5.1
duckdb==1.0.0
confluent-kafka==2.5.3
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==16.1.0
psycopg2-binary==2.9.9
numpy==2.0.1
click==8.1.7
uvloop==0.20.0; platform_system != 'Windows'


